requirements.txt
Django==1.7.6
django-appconf==1.0.1
django-staticfiles==1.2.1
django-tinymce==1.5.3
six==1.9.0
